+++ riverpod_craft/lib/src/data_provider/async_state/async_state.dart
@@ -47,6 +47,73 @@ extension AsynchronousStateExtension<T, ArgT extends Record>
     _ => null,
   };
 
+  R map<R>({
+    required R Function(AsynchronousState<T, ArgT> state) init,
+    required R Function(AsynchronousState<T, ArgT> state) loading,
+    required R Function(AsynchronousState<T, ArgT> state) data,
+    required R Function(AsynchronousState<T, ArgT> state) error,
+  }) {
+    return switch (this) {
+      DataLoading<T>() && final s => loading(s),
+      DataSuccess<T>() && final s => data(s),
+      DataError<T>() && final s => error(s),
+      CommandInit<T, ArgT>() && final s => init(s),
+      CommandLoading<T, ArgT>() && final s => loading(s),
+      CommandData<T, ArgT>() && final s => data(s),
+      CommandError<T, ArgT>() && final s => error(s),
+      ArgCommandInit<T, ArgT>() && final s => init(s),
+      ArgCommandLoading<T, ArgT>() && final s => loading(s),
+      ArgCommandData<T, ArgT>() && final s => data(s),
+      ArgCommandError<T, ArgT>() && final s => error(s),
+      _ => throw UnimplementedError(),
+    };
+  }
+
+  R maybeMap<R>({
+    R Function(AsynchronousState<T, ArgT> state)? init,
+    R Function(AsynchronousState<T, ArgT> state)? loading,
+    R Function(AsynchronousState<T, ArgT> state)? data,
+    R Function(AsynchronousState<T, ArgT> state)? error,
+    required R Function() orElse,
+  }) {
+    return switch (this) {
+      DataLoading<T>() && final s => loading != null ? loading(s) : orElse(),
+      DataSuccess<T>() && final s => data != null ? data(s) : orElse(),
+      DataError<T>() && final s => error != null ? error(s) : orElse(),
+      CommandInit<T, ArgT>() && final s => init != null ? init(s) : orElse(),
+      CommandLoading<T, ArgT>() && final s => loading != null ? loading(s) : orElse(),
+      CommandData<T, ArgT>() && final s => data != null ? data(s) : orElse(),
+      CommandError<T, ArgT>() && final s => error != null ? error(s) : orElse(),
+      ArgCommandInit<T, ArgT>() && final s => init != null ? init(s) : orElse(),
+      ArgCommandLoading<T, ArgT>() && final s => loading != null ? loading(s) : orElse(),
+      ArgCommandData<T, ArgT>() && final s => data != null ? data(s) : orElse(),
+      ArgCommandError<T, ArgT>() && final s => error != null ? error(s) : orElse(),
+      _ => orElse(),
+    };
+  }
+
+  R? mapOrNull<R>({
+    R Function(AsynchronousState<T, ArgT> state)? init,
+    R Function(AsynchronousState<T, ArgT> state)? loading,
+    R Function(AsynchronousState<T, ArgT> state)? data,
+    R Function(AsynchronousState<T, ArgT> state)? error,
+  }) {
+    return switch (this) {
+      DataLoading<T>() && final s => loading?.call(s),
+      DataSuccess<T>() && final s => data?.call(s),
+      DataError<T>() && final s => error?.call(s),
+      CommandInit<T, ArgT>() && final s => init?.call(s),
+      CommandLoading<T, ArgT>() && final s => loading?.call(s),
+      CommandData<T, ArgT>() && final s => data?.call(s),
+      CommandError<T, ArgT>() && final s => error?.call(s),
+      ArgCommandInit<T, ArgT>() && final s => init?.call(s),
+      ArgCommandLoading<T, ArgT>() && final s => loading?.call(s),
+      ArgCommandData<T, ArgT>() && final s => data?.call(s),
+      ArgCommandError<T, ArgT>() && final s => error?.call(s),
+      _ => null,
+    };
+  }
+
   R when<R>({
     required R Function() init,
     required R Function(ArgT? arg) loading,